+++ conanfile.py
@@ -22,14 +22,14 @@ class DnsLibsConan(ConanFile):
     generators = "cmake"
 
     def requirements(self):
-        self.requires("libcurl/7.85.0-adguard4@AdguardTeam/NativeLibsCommon")
+        self.requires("libcurl/7.85.0-adguard5@AdguardTeam/NativeLibsCommon")
         self.requires("libevent/2.1.11@AdguardTeam/NativeLibsCommon")
         self.requires("libsodium/1.0.18@AdguardTeam/NativeLibsCommon")
         self.requires("libuv/1.41.0@AdguardTeam/NativeLibsCommon")
         self.requires("klib/2021-04-06@AdguardTeam/NativeLibsCommon")
         self.requires("ldns/2021-03-29@AdguardTeam/NativeLibsCommon")
         self.requires("magic_enum/0.7.3")
-        self.requires("native_libs_common/2.0.43@AdguardTeam/NativeLibsCommon")
+        self.requires("native_libs_common/2.0.51@AdguardTeam/NativeLibsCommon")
         self.requires("ngtcp2/0.9.0@AdguardTeam/NativeLibsCommon")
         self.requires("pcre2/10.37@AdguardTeam/NativeLibsCommon")
         self.requires("tldregistry/2022-12-26@AdguardTeam/NativeLibsCommon")