+++ net/socket_factory.cpp
@@ -143,12 +143,12 @@ SocketFactory::SocketPtr SocketFactory::make_secured_socket(
 Error<SocketError> SocketFactory::prepare_fd(
         evutil_socket_t fd, const SocketAddress &peer, const IfIdVariant &outbound_interface) const {
     if (const uint32_t *if_index = std::get_if<uint32_t>(&outbound_interface)) {
-        return make_error(SocketError::AE_BIND_TO_IF_ERROR,
-                ag::utils::bind_socket_to_if(fd, peer.c_sockaddr()->sa_family, *if_index));
+        auto err = ag::utils::bind_socket_to_if(fd, peer.c_sockaddr()->sa_family, *if_index);
+        return (err == nullptr) ? nullptr : make_error(SocketError::AE_BIND_TO_IF_ERROR, err);
     }
     if (const std::string *if_name = std::get_if<std::string>(&outbound_interface)) {
-        return make_error(SocketError::AE_BIND_TO_IF_ERROR,
-                ag::utils::bind_socket_to_if(fd, peer.c_sockaddr()->sa_family, if_name->c_str()));
+        auto err = ag::utils::bind_socket_to_if(fd, peer.c_sockaddr()->sa_family, if_name->c_str());
+        return (err == nullptr) ? nullptr : make_error(SocketError::AE_BIND_TO_IF_ERROR, err);
     }
     if (m_router == nullptr) {
         return {};